+++ lib/minidusen/parser.rb
@@ -4,7 +4,7 @@ module Minidusen
     class CannotParse < StandardError; end
 
     TEXT_QUERY = /(?:(\-)?"([^"]+)"|(\-)?([\S]+))/
-    FIELD_QUERY = /(?:\s|^|(\-))(\w+)\:#{TEXT_QUERY}/
+    FIELD_QUERY = /(?:\s|^|(\-))(\w+):(?!:)#{TEXT_QUERY}/
 
     class << self